🧪 TEST: updated tests to handle lowercase desc
